+++ frontend/regional-mma2/src/Components/FighterList/Prospects/Prospects.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import Prospect from './Prospect/Prospect';
+import './Prospects.css';
+
+
+const Prospects=(props)=>{
+    return(
+        <div className="Prospects__Wrapper">
+            <div className="Prospects__Content">
+                {props.fighters.map(fighter=><Prospect key={fighter.name} image={fighter.image} name={fighter.name}/>)}
+            </div>
+        </div>
+    )
+}
+
+export default Prospects;
